refactor(bridge) : 클래스 CalculateImp -> QuotientCalculate 변경

+++ src/main/java/bridge/Calculator.java
@@ -19,7 +19,7 @@ public class Calculator {
         return calculate.mul(a, b);
     }
 
-    public double div(int a, int b) {
+    public int div(int a, int b) {
         return calculate.div(a, b);
     }
